refactor: share admin-style pagination
