+++ src/cook-web/src/components/billing/BillingRecentActivitySection.test.tsx
@@ -160,7 +160,7 @@ describe('BillingRecentActivitySection', () => {
         'module.billing.ledger.usageScene.debug - Debug Course 1 - 15811237246',
       ),
     ).toBeInTheDocument();
-    expect(await screen.findByText(/Apr 6, 2026,/)).toBeInTheDocument();
+    expect(await screen.findAllByText(/Apr 6, 2026,/)).not.toHaveLength(0);
     expect(await screen.findByText('-2.50')).toBeInTheDocument();
     expect(
       screen.queryByText('module.billing.orders.title'),
